fillding with end-of-lines to try to trick github into displaying list of file correctly - failed
if each line is consecutive in the list of files (i.e. no blank line between), github changes it to one long string on one line.
if each line containing a file name is separated by a blank line, github displays it "as is" with the blank lines  - but is also not what I want.
github is stupid in this particular issue.
The problem is made worse because github fails to update the READ.me from my desktop repo when I add, commit and push. stupid.